+++ src/components/image/index.ts
@@ -1,0 +1,4 @@
+import createComponent from '~/utils/createComponent'
+import Image from './image'
+
+export default createComponent(Image)